+++ firebase-perf/src/test/java/com/google/firebase/perf/config/ConfigResolverTest.java
@@ -498,6 +498,83 @@ public class ConfigResolverTest extends FirebasePerformanceTestBase {
   }
 
   @Test
+  public void
+      getIsPerformanceCollectionConfigValueAvailable_noDeviceCacheNoRemoteConfig_returnsFalse() {
+    when(mockDeviceCacheManager.getBoolean(FIREBASE_PERFORMANCE_COLLECTION_ENABLED_CACHE_KEY))
+        .thenReturn(Optional.absent());
+    when(mockRemoteConfigManager.isLastFetchFailed()).thenReturn(false);
+    when(mockRemoteConfigManager.getBoolean(FIREBASE_PERFORMANCE_SDK_ENABLED_FRC_KEY))
+        .thenReturn(Optional.absent());
+    assertThat(testConfigResolver.isCollectionEnabledConfigValueAvailable()).isFalse();
+  }
+
+  @Test
+  public void
+      getIsPerformanceCollectionConfigValueAvailable_noDeviceCacheHasRemoteConfigValueFalse_returnsTrue() {
+    when(mockDeviceCacheManager.getBoolean(FIREBASE_PERFORMANCE_COLLECTION_ENABLED_CACHE_KEY))
+        .thenReturn(Optional.absent());
+    when(mockRemoteConfigManager.isLastFetchFailed()).thenReturn(false);
+    when(mockRemoteConfigManager.getBoolean(FIREBASE_PERFORMANCE_SDK_ENABLED_FRC_KEY))
+        .thenReturn(Optional.of(false));
+    assertThat(testConfigResolver.isCollectionEnabledConfigValueAvailable()).isTrue();
+  }
+
+  @Test
+  public void
+      getIsPerformanceCollectionConfigValueAvailable_HasDeviceCacheNoRemoteConfigValue_returnsTrue() {
+    when(mockDeviceCacheManager.getBoolean(FIREBASE_PERFORMANCE_COLLECTION_ENABLED_CACHE_KEY))
+        .thenReturn(Optional.of(false));
+    when(mockRemoteConfigManager.isLastFetchFailed()).thenReturn(false);
+    when(mockRemoteConfigManager.getBoolean(FIREBASE_PERFORMANCE_SDK_ENABLED_FRC_KEY))
+        .thenReturn(Optional.absent());
+    assertThat(testConfigResolver.isCollectionEnabledConfigValueAvailable()).isTrue();
+  }
+
+  @Test
+  public void
+      getIsPerformanceCollectionConfigValueAvailable_HasDeviceCacheFalseHasRemoteConfigValueFalse_returnsTrue() {
+    when(mockDeviceCacheManager.getBoolean(FIREBASE_PERFORMANCE_COLLECTION_ENABLED_CACHE_KEY))
+        .thenReturn(Optional.of(false));
+    when(mockRemoteConfigManager.isLastFetchFailed()).thenReturn(false);
+    when(mockRemoteConfigManager.getBoolean(FIREBASE_PERFORMANCE_SDK_ENABLED_FRC_KEY))
+        .thenReturn(Optional.of(false));
+    assertThat(testConfigResolver.isCollectionEnabledConfigValueAvailable()).isTrue();
+  }
+
+  @Test
+  public void
+      getIsPerformanceCollectionConfigValueAvailable_noDeviceCacheHasRemoteConfigValueTrue_returnsTrue() {
+    when(mockDeviceCacheManager.getBoolean(FIREBASE_PERFORMANCE_COLLECTION_ENABLED_CACHE_KEY))
+        .thenReturn(Optional.of(false));
+    when(mockRemoteConfigManager.isLastFetchFailed()).thenReturn(false);
+    when(mockRemoteConfigManager.getBoolean(FIREBASE_PERFORMANCE_SDK_ENABLED_FRC_KEY))
+        .thenReturn(Optional.of(true));
+    assertThat(testConfigResolver.isCollectionEnabledConfigValueAvailable()).isTrue();
+  }
+
+  @Test
+  public void
+      getIsPerformanceCollectionConfigValueAvailable_hasDeviceCacheHasRemoteConfigValueFalse_returnsTrue() {
+    when(mockDeviceCacheManager.getBoolean(FIREBASE_PERFORMANCE_COLLECTION_ENABLED_CACHE_KEY))
+        .thenReturn(Optional.of(true));
+    when(mockRemoteConfigManager.isLastFetchFailed()).thenReturn(false);
+    when(mockRemoteConfigManager.getBoolean(FIREBASE_PERFORMANCE_SDK_ENABLED_FRC_KEY))
+        .thenReturn(Optional.of(false));
+    assertThat(testConfigResolver.isCollectionEnabledConfigValueAvailable()).isTrue();
+  }
+
+  @Test
+  public void
+      getIsPerformanceCollectionConfigValueAvailable_hasDeviceCacheHasRemoteConfig_returnsTrue() {
+    when(mockDeviceCacheManager.getBoolean(FIREBASE_PERFORMANCE_COLLECTION_ENABLED_CACHE_KEY))
+        .thenReturn(Optional.of(true));
+    when(mockRemoteConfigManager.isLastFetchFailed()).thenReturn(false);
+    when(mockRemoteConfigManager.getBoolean(FIREBASE_PERFORMANCE_SDK_ENABLED_FRC_KEY))
+        .thenReturn(Optional.of(true));
+    assertThat(testConfigResolver.isCollectionEnabledConfigValueAvailable()).isTrue();
+  }
+
+  @Test
   public void getIsPerformanceCollectionEnabled_notDeviceCacheOrMetadata_returnsNull() {
     when(mockDeviceCacheManager.getBoolean(FIREBASE_PERFORMANCE_COLLECTION_ENABLED_CACHE_KEY))
         .thenReturn(Optional.absent());